State now properly displays images
