+++ clifscatalogv2/app/components/HeroSection.tsx
@@ -3,53 +3,34 @@ import Image from "next/image";
 import Chat from "../functional_components/Chat";
 import "../globalStyles.css";
 import HeroSectionStyles from "../modular_css/HeroSection.module.css";
-import { useState } from "react";
+import { Suspense, useState } from "react";
 import { Session } from "next-auth";
 import SignIn from "../signin/page";
 import { signOut } from "next-auth/react";
 import UserFavoritesDisplay from "../functional_components/UserFavoritesDisplay";
+import SousChef from "../functional_components/SousChef";
 
 interface Props {
   session: Session;
 }
 
 export const HeroSection = (props: Props) => {
-  const [showChat, setShowChat] = useState(false);
-
-  const handleClick = () => {
-    setShowChat(!showChat);
-  };
 
   return (
     <main className={HeroSectionStyles.container}>
       {props.session ? (
         <div>
           <h2>Welcome, {props.session.user.name}</h2>
+          {/* <Suspense> */}
+
           <UserFavoritesDisplay session={props.session} />
+          {/* </Suspense> */}
           <button onClick={() => signOut()}>Sign out</button>
         </div>
       ) : (
         <SignIn />
       )}
-      {!showChat && (
-        <Image
-          src="/chef_kiss.svg"
-          width={100}
-          height={100}
-          alt="emoji of a chef"
-          onClick={handleClick}
-          className={HeroSectionStyles.chatbotImage}
-        />
-      )}
-      {showChat && (
-        <div className={HeroSectionStyles.chatbotDiv}>
-          <div className={HeroSectionStyles.closeSectionDiv}>
-            <button onClick={handleClick}>❌</button>
-            <h3>My Sous-chef</h3>
-          </div>
-          <Chat />
-        </div>
-      )}
+      <SousChef/>
     </main>
   );
 };
