+++ clifscatalogv2/app/components/Navigation.tsx
@@ -1,19 +1,9 @@
-"use client";
 import Image from "next/image";
 import NavigationStyles from "../modular_css/Navigation.module.css";
 import "../globalStyles.css";
-import { useState } from "react";
 import Link from "next/link";
-
+import NavigationThemer from "../functional_components/NavigationThemer";
 export const Navigation = () => {
-  //state for navbar icon theme toggle
-  const [themeClicked, setThemeClicked] = useState(false);
-
-  //click handler for changing navbar icon for theme toggle
-  const handleToggle = () => {
-    setThemeClicked(!themeClicked);
-    console.log(themeClicked);
-  };
 
   return (
     <section className={NavigationStyles.container}>
@@ -39,9 +29,7 @@ export const Navigation = () => {
         <li className={NavigationStyles.navItem}>
           <Link href="#search-recipes">My Sous-chef</Link>
         </li>
-        <button className={NavigationStyles.themer} onClick={handleToggle}>
-          {themeClicked ? "⛅" : "🌛"}
-        </button>
+        <NavigationThemer/>
       </ul>
     </section>
   );
